[routes/teachers.py]: CRUD is done

+++ app/database/models/teacher.py
@@ -6,7 +6,7 @@ class TeacherBase(SQLModel):
     lessons_amount: int = Field(index=True)
     grade_point_avg: float = Field(index=True)
     teacher_courses_amount: int = Field(index=True)
-    positive_rewiews_amount: float = Field(index=True)
+    positive_rewiews_amount: int = Field(index=True)
     
 class Teacher(TeacherBase, table=True):
     id: int | None = Field(default=None, primary_key=True)
@@ -20,7 +20,7 @@ class TeacherUpdate(TeacherBase):
     lessons_amount: int | None = None
     grade_point_avg: float | None = None
     teacher_courses_amount: int | None = None
-    positive_rewiews_amount: float | None = None
+    positive_rewiews_amount: int | None = None
         
 class TeacherPublic(TeacherBase):
     id: int
